Adds HTTP proxy configuration and centralizes HTTP client construction
